requirements.txt
# GuardianAI - Website QA Platform

# =============================================================================
# CORE FRAMEWORK & WEB
# =============================================================================
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-Limiter==3.5.0
Werkzeug==3.0.1
WTForms==3.1.1

# =============================================================================
# DATABASE & ORM
# =============================================================================
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# =============================================================================
# AI & ANALYTICS
# =============================================================================
cohere==4.38
pandas==2.1.3
openpyxl==3.1.5

# =============================================================================
# BROWSER AUTOMATION & WEB SCRAPING
# =============================================================================
playwright==1.40.0

# =============================================================================
# AUTHENTICATION & SECURITY
# =============================================================================
pyotp==2.9.0
qrcode==7.4.2
Pillow==10.1.0

# =============================================================================
# CACHING & JOB QUEUE
# =============================================================================
redis==5.0.1
rq==1.14.1

# =============================================================================
# UTILITIES & HELPERS
# =============================================================================
python-dotenv==1.0.0
markdown==3.5.1
cryptography==41.0.7
requests==2.31.0

# =============================================================================
# LOGGING & MONITORING
# =============================================================================
python-logging-loki==0.3.1

# =============================================================================
# DEVELOPMENT & TESTING
# =============================================================================
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.12.0
flake8==6.1.0
isort==5.13.2